refactor: organise shared utility functions into 'common' directory
