+++ src/lib.rs
@@ -6,7 +6,7 @@ mod pb;
 mod rpc;
 mod store_key_manager;
 mod types;
-mod utils;
+mod common;
 
 pub use modules::*;
